Add (failing) ATDD spec for unlinking cases
This commit adds an acceptance spec for 'unlinking' a case - the
pass condition is sending a correctly-formed unlink request to the
court-data-adapter.

The unlink request is based on reading the adapter code - it surprised
me a little that it was this shape.

Notably, this spec is full of quality/clarity issues I want to
think/talk about, but I felt speed was more important:
- I'm using a flat JSON file for the API response that hides away a
lot of the detail of the interaction
- I'm talking about things like MAAT and ASN instead of things naive
readers could understand
- I'm replicating stubbing that's in `support/webmock.rb`, which I'd
like to combine and use rspec's "shared context" pattern for