Unlink a case from MAAT  (#163)
This commit wires up the 'unlink' button on cases, causing a 
request to be sent to the adapter. As of this commit, this request has 
default values rather than specific ones, as some values are 
necessary to pass the adapter's validation.

The unlink request is a surprising shape - it's not just the reverse of 
the linking request. I've used a non-RESTful route to try and draw 
attention to this.

The associated acceptance spec has some issues: 
- a flat JSON file hides away a lot of the detail of the adapter
- heavy acronym use makes things less clear to a naive reader
- I'm replicating stubbing that's in `support/webmock.rb`, which I'd 
like to combine and use rspec's "shared context" pattern for