Move queryable mixin to acts_as_resource
The name did not reflect functionality
and the mixin is to be used to add custom
court data adaptor related functionlality
to both query objects and resources themselves.

In particular to share `refresh_token_if_required!`
functionality to ensure create and update
calls to adaptor are not using an expired token.