Unlink with unique username
Previously a generic username was sent as part of the unlinking.
However this is not suffiicent as it doesnt allow identification of the
user who unlinked the case. There is a 10 char limit emforced by MLRA
for the username which does not allow for using the user email as the
username. Therefore, a user id will be used instead and VCD will be the
source of truth for identifying the user.