Add user validations
Add missing validations: presence of first and last
names because they are not null on the DB.

Add unique usernames validation. already done on DB

Add format validation: prevent emails being using
as a username to prevent locking issues. and enforce
a bare minimum of logic inline with MLRA/MAAT username

see [devise considerations](https://github.com/heartcombo/devise/wiki/How-To:-Allow-users-to-sign-in-using-their-username-or-email-address#for-activerecord)