Add unlink reasons
add table, model, seed file, factories and
specs for unlink_reasons.

These are used to store unlink reason codes that
users can select when unlinking a defendant -
and are sent to the adaptor. The adaptor, in turn,
sends this to MLRA/MAATDB api which uses them to audit
the reason for unlinking a defendant.

Eventually it is intended that these values are
not stored/persisted by VCD and, instead, they
are retrieved from the adaptor (that, in turn, retrives
them from MLRA/MAATDB-API).