Change unlink method to a patch
A RESTful approach to unlinking

Note: The adaptor will not corece a json string
of an integer into an integer since json
is capable of storing both string and integer.
We should there avoid problems by sending an
integer for the reason code.

see [dry-validation json schemas](https://dry-rb.org/gems/dry-validation/1.5/schemas/#defining-a-schema-with-json-coercion)