Show hearings data currently available via adaptor
Display hearing summaries with links
to hearings Add a hearing page that displays
attendees and hearing events.

currently a hearing is over several days and all
events for any day are shown. This can be handled
VCD side but ideally the adaptor should allow us
to specify a hearing_day for a hearing

e.g. ...v1/hearings/a-hearing-uuid?hearing_day=<01-01-2001>&include-hearing_events