Use prosecution case endpoint for defendant details
The mode of trial decision (aka allocation decision)
is not currently includable by the GET defendant
endpoint of the adaptor. It is only available on the
GET prosecution_case endpoint.

To expose the mode of trial reason we therefore
need to retrieve the entire case and filter
by defendant id.