Add username to link request params
We should be sending the username of the
person requesting the link. This is then
recorded in MAAT API.

Also, modify request spec to use stubs instead
of cassettes as cassettes are proving complex,
difficult to maintain due to churn and obfuscate
the actual responses being received since there appears
to be a utf-8 encoding problem which requires
use of VCR preserve_exact_body_bytes config.