Explicitly define relations and properties for prosecution case
For clarity, simpler testing and avoidance
of NameError exceptions in app we explictly define the
hearings and defendants relations and the
prosecution_case_reference property.

Because the object is created only with properties
and relations defined by the JSON returned from the adaptor
its properties and relation may not actually exist. This
defines them explictly so they can behave more like
an activerecord model object.