Handle decoration of nil
To avoid having to use conditional logic in views:

```
decorated_object = decorate(object) if object
```

A nil object will be decorated with NilClassDecorator
which delegates the basic ruby and rails nil checking
methods (nil?, blank?, present?).

These methods are for use in the `decorate` helper method
only to either return nil or NOT yield. This means a decorated
object in a view will behave as `nil` if `nil` was passed to
the decorator.

```
  - decorated_object = decorate(object)
  if decorated_object && decorated_object.check?
    ...

  OR
  if decorated_object&.check?
    ...
```

```
  decorate(object) do |decorated_object|
    = decorated_object.attribute1
    ...
```

NOTE: The BaseDecorator inherits from SimpleDelegator
so it delegates missing methods to the object
being decorated. However, for the `nil?`
method itself does not seem to get delegated unless
specifically delegated here?!