Handle multiple "cracks"
Since cracks can relate to a specific defendant it is possible,
albeit unlikely (1% of cases have 2+ defendants??!), that there
could be multiple case level cracks. In such cases the view should
be adding a row to the same table, not creating a new table.

Currently the data for cracked_ineffective_trial object does not
include a defendant Id and is per hearing. This means we cannot
identify if a cracked result relates to a specific defendant.
Common Platform would need to change before this can be remedied.