Bump bootsnap 1.6.0 --> 1.7.0
Getting build errors as below. bootsnap is
the dependency that use msgpack and it has released
a fix today thankfully.

```
Your bundle is locked to msgpack (1.4.1-java), but that version could not be
found in any of the sources listed in your Gemfile. If you haven't changed
sources, that means the author of msgpack (1.4.1-java) has removed it. You'll
need to update your bundle to a version other than msgpack (1.4.1-java) that
hasn't been removed in order to install.
```