Extract hearing "listing" information to partial
Continuation of extraction of logical view segments
to partials for isolation of funtionality and for testing
purposes.

The "listing" information is a hearing's fundamental
information for when the cases hearing was entered
into the court record to be heard - hearing type, court
name and date/time it is to start.