Explicitly define required properties
These properties of the defendant
are essential to the correct functioning
of the UI and therefore should be defined
for clarity.

The app requires certain properties to be
either readable or read/writable. The property
definition itself (from json_api_client gem)
will create accessors, but the app only needs
certain attributes the be assignable/writable.