Refactor the prosecution case "show" flow
There can only be one prosecution_case
per URN - we might want to consider raising
an error if there is ever more than one.

This PR therefore removes the need to iterate
over an array or results to retrieve the
prosectution_case object.

This removes alot of clutter and logic
from the views. It also moves the decoration
of the prosecution_case to the controller so
any views/partials do not need to decorate
individually.