Configure vcr hook to match on method, path and query
Allows cassettes recorded against one
adaptor domain (or local) to function
as expected for feature tests, irrespective
of the app configuration which defines which
adaptor domain to make requests to.

added method matching too to mitigate against
the unlikely scenario that a cassette has a
POST to the same path and query.