Disable rubocop Rails/HelperInstanceVariable
I do not consider this to be relevant in the context
of a class tha is used to create an object for use in
views.

I may raise an issue with rubocop. I note that memoized
instance variables in helpers do not violate this cop
due to an issue that was accepted for it.