Configure and use low-level caching in development
Of the various methods of caching available in rails
this is the only one that does what we need - namely
caching the response from the adapter and significantly
speeding up rendering.

This works fine in development but will require
a SEPARATE redis server to act as a cache store
in production.