Only query prosecution_case endpoint
Ideally we should be hitting the hearing endpoint only, however, since pagination
currently requires having knowledge of all hearings via the prosecution case endpoint
it is pointless and time-consuming to query both endpoints when the prosecution case
endpoint contains all the info we need.

As a further optimisation we could pass a "pagination collection" instead
and hit the hearing endpoint alone.