Remove cracked/vacted on link
This link depended on older mechanism
for querying a hearing. It was also a simple
drop in that was not part of designs or user-research

The introduction of pagination (and later sorting)
has broken this simplistic functionality. To get
it working with pagination (and sorting) would have
involved tigh coupling and considerable extra processing
which a) seems excessive for the benefit b) should
be worked as as a separate piece of work if needed.