Validate the minimum of visible chars for searching
The defendant name search on Common platform
allows any number charaters. However this
could result in a large payload response
that causes a timeout during the request
response cycle. This causes a 500 error
handler to kick in and display the
"Something went wrong" page.

To mitigate this we can ensure no queries
with a search term of 1 (non-whitespace) char
is submitted to the adaptor.

This is pending a "proper" solution
that would require pagination to exist on the
common platform API and the Court data adaptor