Add Cypress orb to CircleCI to run tests
Cypress was unable to run because the binaries were missing. The Cypress
orb is added to handle the install and run of tests. This removes
previous implementation where we run the Rails server and tests via yarn
commands, and instead handles this in /.circleci/config.yml with Cypress
orb and wait-on module.