Clean up Dependabot config
Remove the ignoring of certain packages and their versions because the reasons for this are out of date as the packages are now ahead in versions.

The Dependabot minor versions were ignored during the migration of dependabot to GitHub-native Dependabot
These ignore versions are out of date as the gems are now ahead in versions.