Add an initial CODEOWNERS file (#782)
Add code owners ministryofjustice/laa-assess-a-claim team to * all changes to the repo

Codeowners is a github functionality. It allows for automatically adding a reviewer to a PR if it changes something they "own" e.g.
any PR that changes a javascript file will request a review from the frontend team members etc.