Move Puma to single mode
This is because we are currently running a single worker in Puma's cluster mode,
which means we could just opt for the single mode (this means no cluster manager/runner needed)

This should prove to be minor local development performance improvements.

Also, removes the log WARNING: Detected running cluster mode with 1 worker.