AAC-379 - Fix mailer errors in local development environment  (#824)
* Add config to set default host for mailers

When generating a devise email after a user is created etc, an error complains
that the host is not set. Whilst there is a default host for the app,
unlike controllers, mailers do not have the same context and there isn't
a default, so this needs to be set explicitly. The config added sets the host
to be consistent with the default Rails host.

The host is required for the devise mailer because it calls Rails url
helpers to generate urls.

* Change default_url_options to be consistent with test env

Co-authored-by: Paul King <84575729+pking-moj@users.noreply.github.com>