Add test for errors and removes former error handling
ConnectionErrors were previously caught in SearchesController, this does
not catch some of the errors we saw in LaaReferencesController or
DefendantsController. The error was displayed in plain text and not very
easily visible by the user either, it is now displayed in a flash
message at the top of the page.