AAC-403 Add 500 error handling (#832)
* Add 500 error handler to json-api-client wip

* Add a new exception for handling CDA 500 responses

Add a InternalServerError exception to handle CDA 500 http responses.

We "assume" HTTP 500 responses have a "body attribute" but it may be nil, which arrives as an empty string.

Add a test to check that we raise exceptions for 500 http responses from CDA.

* Capture Net::HTTP timeouts and render on own page

Create a /504 route to direct any connection errors to e.g. Net::Http timeout errors or
if needed any 504 http responses we get back upstream.

Create a view and test the new errors controller path.

Send this caught Net::Http exception to Sentry before rendering a specialised /504 view
that contains clearer user information about what's going on.

* Update /504 error feature test and error text

Internal server error handling on json-api-client level is removed and
replaced by ApplicationController level error handling of
Net::ReadTimeout. Error feature tests are also added and error text for
/504 is provided.

* Remove unneeded base_spec for json-api-client

* Fix i18n linting complaint

* Add flash alert for Net::ReadTimeout error, fix case statements

* Remove unneeded error title

* Remove connection error related error handling

* Add test for errors and removes former error handling

ConnectionErrors were previously caught in SearchesController, this does
not catch some of the errors we saw in LaaReferencesController or
DefendantsController. The error was displayed in plain text and not very
easily visible by the user either, it is now displayed in a flash
message at the top of the page.

* Update error testing in searches request specs

* Update error wording with helpdesk number

Co-authored-by: Obsiye <mrabdirizak@gmail.com>