AAC-162 Cypress Test Cookies and Login/Logout (#836)
* Update yarn dependency

This is because a log was presented when running yarn scripts:

Running:  index/sign_in.spec.js                                                           (1 of 1)
Browserslist: caniuse-lite is outdated. Please run:
npx browserslist@latest --update-db

Why you should do it regularly:
https://github.com/browserslist/browserslist#browsers-data-updating

* Test logout with cypress

Test the sign out process with cypress.

* Add cypress tests for cookie banner and page

Add cypress tests for cookie banner updates.

Add cypress test for alteration of cookie on its settings page.

* Block Google  analytics for Cypress

Block google analytics for Cypress to prevent corruption of data in Google Analytics.

* Fix javascript styling to standardjs recommendations

Fix javascript styling to pass linting.

* Add data attributes for cypress testing of cookies

Add data attributes for cookie banner buttons: accept, reject and settings of cookies

Add data attribute for form submit button on the cookie settings page.

* Add cypress tests for hiding cookie banner

Add data attribute to the hide message button in the cookie banner.

Test the hide message functions as expected.

* Add cypress test for logging out

Should not see the search filters page when logging out.

* Add cypress tests for checking cookies

Add tests to check for the cookies we store.

Add test to check that turning the cookies off in the cookies settings page is possible.

Add a command to make these cookies check less wordy and easy to change.

* Amend cypress cookie testing

Amend the cypress tests to match the below logic in the app:

By default analytics_cookies_set cookie is set and its value is false.
By default cookies_preferences_set cookie is not set until a choice is made by the user around cookies.

* Refactor cypress test with variables

Add a variable to pull out text into a single place to be easily changeable.

match case set to false because "change your cookie settings' is a link with no capital letters on the banner
aswell as being a title on the cookie settings page.

* Refactor cookie cypress tests

Seperate the testing of the cookie banner from the settings page into seperate files
to be easier to read.

Refactor code to move value text into constants that can be easier to change and DRY.

* Refactor cypress tests

Clean up javascript code. Remove unnecessary string literals / template strings.

* Explicity describe cypress comman envbanne

Change cypress custom command for checking the banner regarding the environment.
This is because the cookie banner is seperate to the banner containing the description of the websites environment etc.

* Refactor cypress tests for variables

Move variables closer to top of file to make it easier to read