Update search_model for CDAPI/V2 calls
Add a feature flag called  DEFENDANTS_SEARCH. If true, make search calls to cdapi
else call cda v1 as normal.

Abstract the call to CdApi::SearchService and call it.

Handle exceptions raised by calling CdApi::SearchService. Always return an empty collection if there's an exeption caught.
This is so that it's consistent with the CDA v1 empty results.

Log information about exceptions to improve searching through logs.

Update search_spec.rb to test when the DEFENDANTS_SEARCH flag is true and that it calls CDAPI