Remove court_data_api_resource module
Remove this file because the fix was to add the test url in the .env.test file. This is to
set the environment variable Rails.configuration.x.court_data_api_config.uri , which is used for ActiveResource calls.

No need to use the module anymore and no need for the method_missing hack.