AAC-548 Implement defendant endpoint (#952)
* Update search_model for CDAPI/V2 calls

Add a feature flag called  DEFENDANTS_SEARCH. If true, make search calls to cdapi
else call cda v1 as normal.

Abstract the call to CdApi::SearchService and call it.

Handle exceptions raised by calling CdApi::SearchService. Always return an empty collection if there's an exeption caught.
This is so that it's consistent with the CDA v1 empty results.

Log information about exceptions to improve searching through logs.

Update search_spec.rb to test when the DEFENDANTS_SEARCH flag is true and that it calls CDAPI

* Update tests env config file

Add CDAPI url as this is required for the activeresource calls to CDAPI.

Add the DEFENDANT_SEARCH environment variable to be used as a feature flag.

* Add CdApi::SearchService

Add the CdApi::SearchService in the folder cd_api (for organisational purposes: differentiates it from CDA)

Add a private urn method to clean up the prosecution case reference (trim and upcase it) before sending the request to CdApi

Call ReferenceParserService to return either a :asn or :nino key and its value. This can be used in a single call CdApi::Defendant.find(:all, params: { reference.kind => reference.value })
with dynamic params for either asn or nino

Test the above

* Seperate webmock for CDAPI and CDA

Seperate out into seperate files the webmock stub requests for CDAPI and CDA.

This is for organisational purposes (less confusion)

Some of the stubs load json files as the response as the json response can be rather large and take up file space.

* Add defendant_body.json for use with stubbing

Add the defendant_body.json, which contains fake json data from CDAPI.

This is so that its used within stub responses in webmock/cd_api

* Remove court_data_api_resource module

Remove this file because the fix was to add the test url in the .env.test file. This is to
set the environment variable Rails.configuration.x.court_data_api_config.uri , which is used for ActiveResource calls.

No need to use the module anymore and no need for the method_missing hack.

* Add Defendant activeresource 'model'

Add a cda_api folder under app/models/ to make it more distinguised from other models.

Move the laa_references and base_model under the app/models/cd_api directory

* Move searches shared_examples test to support folder

Move the searches shared_examples to the support folder to be reused with cd_api/searches_spec.rb

* Update laa_references controller model name

Use CdApi:: module as the LaaReferences that uses ActiveResource is now in the app/models/cd_api directory

* Add searches requests specs for CDAPI calls

Add requests specs for the calls to CDAPI. This is an almost exact duplication of the CDA requests specs.
However, the difference is in the Feature flag being enabled for DEFENDANT_SEARCH on CDAPI instead of CDA

* Add ReferenceParserService for CDAPI defendant search calls

This is an almost exact copy of CDA/ReferenceParser , however :arrest_summons_number is now :asn for CDAPI
and :national_insurance_number is now :nino for CDAPI.

Also the CDA reference_parser is in the lib folder and may entangle these two if used as the sole dynamic referenceparser for both CDAPI and CDA callls

* Small fix for log warnining Deprecation Warning

The fix was to change the file name to export_csv.rb from export.csv.erb as this caused the below log warning

The issue was:

DEPRECATION WARNING: Rendering actions with '.' in the name is deprecated: users/export_csv.erb

* Allow VCR recording of CDAPI calls

Add CDAPI url calls as an allowed uri in the VCR configuration

Remove PII from VCR requests to CDAPI regarding the BASIC authentication

Removed unused :vcr_cud_request

* Add a TODO to test the ASN format

The ASN number format is not tested as the only validations on it are in the Search.rb model
which makes sure it's either a number or letter (a-zA-z)

* Fix test to use _csv.erb not .csv.erb

As the file name has changed to export_csv.erb from export.csv.erb

This is a part of an earlier fix around:

DEPRECATION WARNING: Rendering actions with '.' in the name is deprecated: users/export_csv.erb

* Add feature tests and vcr cassettes for CDAPI search

Add feature tests for CDAPI search defendants requests

Add related VCR cassettes

* Reuse view partial for showing search results

Reuse the resutls/_defendant view partial for CDAPI results

* Update cdapi stub name

This specifies that it's for v2 stubs not v1

* Use symbol as argument to Feature.enabled?

Use :defendants_search in the parameter of Feature.enabled?(:defendants_search)

* Undo fix for log warning on csv.erb partial

Undo the fix as per Pull request review. This work will be a seperate branch with possibly a better solution
and more documentation

* DRY out cd_api/search_service

Clean code by removing duplication and using a dynamic params value

* Remove not needed TODOs

Remove the todo around CDA arrest_summons_number format as this code will be removed in future.

Remove the dob.to_s and the TODO beside it because the ActiveResource using json format in the request
thus converting the dob (Date object) to a string, which is exactly how CDAPI needs it

* Inherit from CDA::...::ReferenceParser

Inherit the ReferenceParser from CDA in CdApi::ReferenceParserService and override the #kind method
, which is different between v1 and v2 (:arrest_summons_number becomes :asn)

* Split out search model spec for v2

Split out the search model spec to models/cdapi/search_spec.rb (v2) and models/search_spec.rb (v1)
so that it can be easier to remove v1 related files in future

* Add logging for V1 and V2 defendants search

This will make it easier to find V1 vs V2 calls in the logs