Add CdApi decorators for V2 ProsecutionCase page
Add DefenceCounselDecorator, which is similar for v1 except the attribute role is now status. Also, build the name from
the attributes: first_name, middle_name and last_name with considerations when some portions are missing.

Add HearingDayDecorator which adds a to_datetime method that converts v2 sitting_day to datetime for use on the view pages.

Add HearingSummaryDecorator to use defence_counsel_list for v2 instead of provider_list in v1. TECH DEBT: There is logic arround attendance days that
needs to be done, to make sure on a multi day hearing, the provider is listed only for the day that they've attended and not all. At the moment, it's all.

Add OverallDefendantDecorator with two public methods, name and linked? (for maat_reference) all in v2

Add a ProsecutionCaseDecorator, which has public methods to support the view pages for prosecution_cases in v2. Major difference between v1 and v2 here is that hearings is from hearing_summaries and defendants is from overall_defendants found in the CDAPI /hearingsummaries response

Pass in the decorator_class when decorating an object because this means we don't have to couple feature flags with CdApi:: decorators or V1 decorators.