Add Techdebt to build name for V2
In V1 name was returned as an attribute from CDA. In V2 hearingsummaries name is portioned out to first_name, middle_name and last_name.

Add Tech debt to create a service with validations to build the name. It means it's more reusable and the must have portions of a name are there e.g.
what if last name isn't there but first name and middle name is etc.