AAC-559 Integrate CDAPI HearingSummaries to Prosecution Case Page (#985)
* Add hearing_summaries feature flag in .env.test

* Add V2 prosecution_case page call to controller

Add hearing_summaries activeresource request to the prosecution case controller
Add the decorator_class argument for the specialised CdApi/V2 ProecutionCase decorator

Use the param[:id] as the urn to build the breadcrumb link to the prosecution case page (This is until we can get the URN back as part of the hearing
summaries endpoint call v2)

Update the search model to take in a version2 param, which decides when to use the version 2 defendants_search. This decouples the Search model from
the version call. Update the version 2 argument to be from defendants search on the searches contorller and hearing_summaries on the prosecution case controller

* Add request test for V2 prosecution case page

Add tests for rendering the prosecution_case/show page
Add tests for exception handling. This sends the user to the search results page - TechDebt: it may change to a more convenient solution for the user

Also, clean up code on the spec/requests/linking/link_defendant_maat_reference_v2_spec.rb

* Add webmocking support for hearingsummaries v2 call

Webmock the hearingsummaries cdapi call and mock the response with a hearingsummaries json body in the spec fixtures directory.

Update the locales for providing an error message when the prosecution case page throws an activeresource error and directs the user
to the search results page

* Update authorisation for prosecution case controller

Allow managers and caseworkers to be authorised to call V2 hearingsummaries on the prosecution case page

Test the ability to authorise CdApi queries

* Update search model to take version2 argument

the version2 argument will be used to set whether or not the search model should call V1 or V2

This is passed in from either the prosecution case controller or the seraches controller at the moment

Test the above

* Add CdApi decorators for V2 ProsecutionCase page

Add DefenceCounselDecorator, which is similar for v1 except the attribute role is now status. Also, build the name from
the attributes: first_name, middle_name and last_name with considerations when some portions are missing.

Add HearingDayDecorator which adds a to_datetime method that converts v2 sitting_day to datetime for use on the view pages.

Add HearingSummaryDecorator to use defence_counsel_list for v2 instead of provider_list in v1. TECH DEBT: There is logic arround attendance days that
needs to be done, to make sure on a multi day hearing, the provider is listed only for the day that they've attended and not all. At the moment, it's all.

Add OverallDefendantDecorator with two public methods, name and linked? (for maat_reference) all in v2

Add a ProsecutionCaseDecorator, which has public methods to support the view pages for prosecution_cases in v2. Major difference between v1 and v2 here is that hearings is from hearing_summaries and defendants is from overall_defendants found in the CDAPI /hearingsummaries response

Pass in the decorator_class when decorating an object because this means we don't have to couple feature flags with CdApi:: decorators or V1 decorators.

* Test V2 decorators for prosecution case page

Use factories to mock objects that the decorators take as an argument

Test when hearing_summaries flag is set to true

Testing is similar to V1 decorators except that above and some method differences

* Add ProsecutionCase model V2 and child models

Add ProsecutionCase model which inherits from ActiveResource
Add HearingSummary model with attr_accessor :day which is used to handle how a HearingSummary object with multiple days is split into seperate records on the table

Add associations to the ProsecutionCase V2 model: overall_defendants and hearing_summmaries
and add class_name for organising folders structure so that cdapi models can be kept seperate from v1 models

* Add factories related to V2 CDAPI models

This is used for testing purposes, removes the need for instantiating objects directly

Also, allows for at least one extra place to have attributes of a CdApi model to be represented other than the model
This means that if you wanted to check what attributes hearing_day had, you could also check the factories

* Add V2 ProsecutionCase Views

Add conditional to the show page to either show v1 or v2 views depending on the hearing_sumamries feature flag.

Change rawresponse view to allow for activeresource deeply nested objects to be represented in a flat way like in v1

* Add view tests for v2 prosecution_case

Add tests related to the views on V2 prosecution case page

* Add feature tests and update support tests

Add feature test for hearing_sorter_v2. there doesn't seem to be one for the whole prosecution case page atm, which is TechDebt.

Update the spec/support for v2 to use factories

Add search model test when v2 hearing_summaries flag is on

Add vcr casette for the above feature tests on hearing_sorter_v2

Add helper method to application_helper so the views can check if v2 hearing summaries feature flag is on

* Ignore rubocop model association warning

This is because rubocop expects the model to be a ActiveRecord when it's ActiveResource, ignore the cd_api/ models
until a solution is found (TECHDEBT)

Clean up prosecution case model

* Rename ProsecutionCase cdapi model to CaseSummary

This is because the CdApi::ProsecutionCase name conflicted with the hearing-to-hearing page which had prosecution case nested in it that was mapped to CdApi::ProsecutionCase
automatically by ActiveResource.

This prevents any conflicts with nested resources of the same name across the app.

Update all tests, files, methods and classes

* Use prosecution_case_reference from cdapi v2 response

Use the prosecution_case_reference that is now returned as part of a call to the V2 hearingsummaries endpoint on CDAPI

Update casettes and json stubs and files to adhere to this

* Add Techdebt to build name for V2

In V1 name was returned as an attribute from CDA. In V2 hearingsummaries name is portioned out to first_name, middle_name and last_name.

Add Tech debt to create a service with validations to build the name. It means it's more reusable and the must have portions of a name are there e.g.
what if last name isn't there but first name and middle name is etc.

* Remove attr_accessor for v2 defendant

* Clean up comment