Use params.fetch to provide default search params
Use params.fetch instead of params.require as this prevents any failure if there are no search params.

One case being, the >> render 'new' << code on the #create route of the searches controller removes the search params leaving the endpoint as /searches
this was an invalid URL because the searches controller requires (params.require) search key in the params. Instead use the params.fetch and provide the current_search_params (most recent search params) from the session store.

Add a before hook in the controller to handle empty params.

Redirect to the search filters new page when there are no params on the url for /searches

Test this with request tests on the searches controller and a feature test on the cookies page.

Add tests for V1 and V2 of the service