Add data-turbo = false to fix flashes failure
Flashes were not appearing after successfully signing in or unsuccessfully.

This was because Devise and Turbolinks don't exactly work well together.

Therefore add a data-attribute of data-turbo to all the submit buttons related to Devise and user related features.
Set it to false, to prevent turbolinks from intercepting the call.