Update the hearing decorator to list out defence counsels
Add a defence_counsels_list public method to create a list of defence counsels e.g. John Smith (QC) for Jane Doe

In order to create the above list, we have to first map the defendants in the defence_counsels.defendants (which is a list of ids) to the
defendant details in the prosecution_cases under the hearing object.

Call the DefenceCounselsListService to build the list of names using a name builder logic as we get parts of a fullname e.g. first_name, last_name

Add map_defence_counsels to map the defendant_ids to the defendant details. It searches through each prosecution case that the hearing is linked to in order to find
the details of the defendant.