Update hearing decorator to skip unmatched defendant ids
When mapping the defendant_ids to defendant_details some defendant_ids may fail to match. Added an if statement
to make sure that the defendant_id is returned when it fails to map instead of a nil value.

Update the defence counsel list service to turn a defendant_id to e.g. John Smith for not avialable . This is because the defendant id failed to map.

If the match passes it becomes a CdApi::Defendant object else it is left as an id : [<CdApi::Defendant...>, '1234-abcde-4567-adsfas', <CdApi::Defendant...>]
becomes ['John Smith (QC) for Jane Doe', 'John Smith (QC) for not avialable', 'John Smith (QC) for Defendant3']