AAC-325 Defendant-Provider relationship on Hearing days (#1168)
* Enable defence counsel to defendent relationship

Map the defence_counsel.defendants (defendant ids) to the hearing_summary.defendants (objects) in the HearingSumamries Decorator.
This allows for the defence counsel summary decorator to easily build a name_status_and_defendant list e.g. John Doe (QC) for Defendant ABC

Add duplicated defence counsels when there are duplicated defendants in DefenceCounselDecorator e.g.
John Doe (QC) for Defendant A
John Doe (QC) for Defendant B
which is preferred over "John Doe (QC) for Defendant A and Defendant B" because of clarity on the case page.

* Add tests for update defence counsel list

Add unit tests to check for how the defence counsel list is built when there are multiple defence counsels with multiple defendants. Aswell as to test when there are no defendants for the defence counsel.

Add unit tests to Defence counsel decorator method name_status_and_defendants , to ensure the providers attending is built like "John Doe (QC) for Defendant A<br>John Doe (QC) for Defendant B<br>Bob Smith (QC)"

* Fix rubocop issues

* Refactor #formatted_multiples_defendant_names

Reduce size of method by removing unnecessary lines

Also, rename formatted_multiples_defendant_names method to formatted_defendant_names

* Add the decorator to a blank hearing

Decorate a blank hearing with the HearingDecorator for consistency

* Update the hearing decorator to list out defence counsels

Add a defence_counsels_list public method to create a list of defence counsels e.g. John Smith (QC) for Jane Doe

In order to create the above list, we have to first map the defendants in the defence_counsels.defendants (which is a list of ids) to the
defendant details in the prosecution_cases under the hearing object.

Call the DefenceCounselsListService to build the list of names using a name builder logic as we get parts of a fullname e.g. first_name, last_name

Add map_defence_counsels to map the defendant_ids to the defendant details. It searches through each prosecution case that the hearing is linked to in order to find
the details of the defendant.

* Add tests for HearingDecorator#defence_counsel_list

Add unit tests to the hearing decorator spec for the defence_counsel_list method

It tests that the defendant_ids are mapped to the defendant details in the hearing's prosecution case

It tests that a call is made to the DefenceCounselsListService

* Update hearing decorator to skip unmatched defendant ids

When mapping the defendant_ids to defendant_details some defendant_ids may fail to match. Added an if statement
to make sure that the defendant_id is returned when it fails to map instead of a nil value.

Update the defence counsel list service to turn a defendant_id to e.g. John Smith for not avialable . This is because the defendant id failed to map.

If the match passes it becomes a CdApi::Defendant object else it is left as an id : [<CdApi::Defendant...>, '1234-abcde-4567-adsfas', <CdApi::Defendant...>]
becomes ['John Smith (QC) for Jane Doe', 'John Smith (QC) for not avialable', 'John Smith (QC) for Defendant3']

* Add defence_counsel_list_service test

Update the defence_counsel_list_service unit tests.

* Add models related to the hearing object

* Update the attendees view to use hearing.defence_counsel_list

Update the attendees view to call the defence_counsel_list on the hearing.hearing / hearing_details object to prepare the list of defence_counsels and their defendants

e.g.

John Smith (QC) for Jane Doe<br>John Smith (QC) for not available<br>Tom Jones (junior) for Micheal Jackson

* Update unit tests for HearingDecorator

Add unit tests for the HearingDecorator

Test that the defendant_ids in the defence_counsel is mapped correctly.
Test that the defendant details are being searched for on all prosecution cases linked to the hearing.

* Add factories for hearing object

Add factories for unit testing the hearing object.

Add a factory for hearing and hearing_details: hearing object has an attribute hearing within it so to prevent duplciation hearing_details factory name was created

Update the cracked_ineffective_trial_decorator to make use of the new hearing factories.

* Update hearing attendees_v2 view tests

Update the attendees view unit tests to make use of the hearing factories

* Fix rubocop failures

Rearrange the models because of loading issue with zeitwerk e.g. module Hearing and Model Hearing can't both exist in the app at the same time.

Update tests

* Fix translation using I18n.t

* Guard against failed defendant match in defence counsels

Return 'not available' as the defendant details when the defence_counsels.defendants fails to match
the hearing_summaries.defendants

e.g. John Smith (QC) for not available

* Add integration test and capitalize defence counsel names

Add integration tests for defence advoctes on the hearing page

Capitalize the name of the defendant and defence counsel as they're built

* Capitalize names of defence counsel list

Capitalize the names of the defence counsels and their defendants e.g.
John Smith (qc) for Jane Doe

* Guard against failed defence_counse.defendant match

If the defendant hasn't been matched, it could be a nil or id. Guard against this so that
it becomes 'John Smith (QC) for not available' for either case.

* Use I18n where it's not loaded

Use I18n.t where the I18n classs is not loaded.

Don't use I18n.t when it's loaded instead just use t(...)

* DRY out code for decorating hearings

* Fix code smells alerted by sonarqube